Trim  MyEditor state flow
