+++ src/pages/PostWrite/index.js
@@ -1,15 +1,11 @@
 import React, { useEffect } from "react";
-import { useSelector, useDispatch } from "react-redux";
+import { useDispatch } from "react-redux";
 
-import { selectors, actions } from "data";
+import { actions } from "data";
 import { MyEditor } from "components";
-import { useEditorState } from "components/MyEditor/hooks";
 
 export default function PostWriteComp() {
   const dispatch = useDispatch();
-  const userSession = useSelector(selectors.user.getUserSession);
-  const { id } = userSession;
-  const [editorState, setEditorState] = useEditorState(id);
 
   useEffect(() => {
     dispatch(actions.editorState.populateEditorState());
@@ -17,12 +13,6 @@ export default function PostWriteComp() {
     return () => {
       dispatch(actions.editorState.resetEditorState());
     };
-  }, []);
-  return (
-    <MyEditor
-      editorState={editorState}
-      setEditorState={setEditorState}
-      id={id}
-    />
-  );
+  }, [dispatch]);
+  return <MyEditor />;
 }
